+++ bigmler/defaults.py
@@ -122,7 +122,8 @@ FLAGS = {'BigMLer': [{'flag': 'debug', 'type': 'boolean'},
                      {'flag': 'balance', 'type': 'boolean'},
                      {'flag': 'weight_field', 'type': 'string'},
                      {'flag': 'objective_weights', 'type': 'string'},
-                     {'flag': 'label_aggregates', 'type': 'string'}]}
+                     {'flag': 'label_aggregates', 'type': 'string'},
+                     {'flag': 'missing_strategy', 'type': 'string'}]}
 
 
 def get_user_defaults(defaults_file=DEFAULTS_FILE):