+++ bigmler/options.py
@@ -35,6 +35,7 @@ def create_parser(defaults={}, constants={}):
 
     max_models = constants.get('MAX_MODELS')
     plurality = constants.get('PLURALITY')
+    last = constants.get('LAST_PREDICTION')
 
     version = pkg_resources.require("BigMLer")[0].version
     version_text = """\
@@ -898,6 +899,17 @@ under the License.""" % version
                         help=("Comma-separated list of aggregation functions "
                               "for the multi-label field labels."
                               " Allowed aggregates: count, first and last"))
+
+    # Strategy used in predictions when a missing value is found for the
+    # field used to split the node.
+    parser.add_argument('--missing-strategy',
+                        action='store',
+                        dest='missing_strategy',
+                        default=defaults.get('missing_strategy', last),
+                        choices=["last", "proportional"],
+                        help="Strategy used when the field used in the rules"
+                             " to next nodes is missing in the input data."
+                             " Allowed values: last or proportional")
 
     # The following options are only useful to deactivate the corresponding
     # oposed default values